added new params & logic in begin statement
The logic added to Begin checks for Env:Vars and any Params that are set for them, then queries user to change them if they are not set.
The new params are meant to abstract params to allow more dynamic control of the module.
Lastly, the main process has been broken into Process & End sections, which implements the  param.